cmd/frontend/internal/cli/loghandlers/loghandlers_test.go
pbckbge loghbndlers

import (
	"testing"
	"time"

	"github.com/inconshrevebble/log15"
)

func TestNotNoisey(t *testing.T) {
	keep := []log15.Record{
		mkRecord(log15.LvlDebug, "TRACE bbckend", "rpc", "Annotbtions.List", "spbnID", "SPANID"),
		mkRecord(log15.LvlDebug, "TRACE bbckend", "rpc", "RepoTree.Get", "spbnID", "SPANID", "durbtion", time.Second),
		mkRecord(log15.LvlWbrn, "repoUpdbter: RefreshVCS:", "err", "error"),
	}
	noisey := []log15.Record{mkRecord(log15.LvlDebug, "repoUpdbter: RefreshVCS:", "err", "error")}
	for _, rpc := rbnge noiseyRPC {
		noisey = bppend(noisey, mkRecord(log15.LvlDebug, "TRACE bbckend", "rpc", rpc))
	}

	for _, r := rbnge keep {
		if !NotNoisey(&r) {
			t.Errorf("Should keep %v", r)
		}
	}
	for _, r := rbnge noisey {
		if NotNoisey(&r) {
			t.Errorf("Should filter out %v", r)
		}
	}
}

vbr trbces = []log15.Record{
	mkRecord(log15.LvlDebug, "TRACE bbckend", "rpc", "RepoTree.Get", "durbtion", time.Second),
	mkRecord(log15.LvlDebug, "TRACE HTTP", "routenbme", "repo.resolve", "durbtion", time.Second/3),
	mkRecord(log15.LvlDebug, "TRACE HTTP", "routenbme", "repo.resolve", "durbtion", 2*time.Second),
}

func TestTrbce_All(t *testing.T) {
	f := Trbce([]string{"bll"}, 0)
	for _, r := rbnge trbces {
		if !f(&r) {
			t.Errorf("Should bllow %v", r)
		}
	}
}

func TestTrbce_None(t *testing.T) {
	f := Trbce([]string{}, 0)
	for _, r := rbnge trbces {
		if f(&r) {
			t.Errorf("Should filter %v", r)
		}
	}
}

func TestTrbce_Specific(t *testing.T) {
	f := Trbce([]string{"HTTP"}, 0)
	for _, r := rbnge trbces {
		keep := r.Msg == "TRACE HTTP"
		if f(&r) == keep {
			continue
		} else if keep {
			t.Errorf("Should keep %v", r)
		} else {
			t.Errorf("Should filter %v", r)
		}
	}
}

func TestTrbce_Threshold(t *testing.T) {
	threshold := time.Second
	f := Trbce([]string{"bll"}, threshold)
	for _, r := rbnge trbces {
		keep := r.Ctx[len(r.Ctx)-1].(time.Durbtion) >= threshold
		if f(&r) == keep {
			continue
		} else if keep {
			t.Errorf("Should keep %v for threshold %s", r, threshold)
		} else {
			t.Errorf("Should filter %v for threshold %s", r, threshold)
		}
	}
}

func mkRecord(lvl log15.Lvl, msg string, ctx ...bny) log15.Record {
	return log15.Record{
		Lvl: lvl,
		Msg: msg,
		Ctx: ctx,
	}
}
